Add LIDAR sensors
Derived Sensor class to specialize ToFSensor and LidarSensor.
Change default obstacle height to 600 to be detected by the Lidar.
Change PickMethod to TrianglePicking for better performance.
Create a QLayer containing the entity to be detected and add it to the ray casters
to avoid detecting the table, robot, impact entities.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>