tools: monitor: add robot 2022
Update robot STEP file to 2022 version.

Also change the way of handling the robot order entity (the one that indicate order position).
For the real robot, we use a detailed STEP file with colors, but it is overkill for robot order
entity (in terms of load time, memory consumption and code complexity).
Using simple shape from a OBJ file is much faster and the code is much simpler since it does
not required to load an entire scene and modify.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>