tools: copilot: add Detector tool to compute dynamic obstacles
The Lidar is now connected to Raspberry Pi instead of STM32.
This commit is only related to the emulator version.

The Detector tool computes dynamic obstacles based on Lidar data and robot
position.
The robot position is provided by mcu-firwmare through Protobuf messages.
In emulator mode (Lidar in not connected and uart_port option of Detector
is set to None), Lidar data are generated by the Monitor, send to the
Copilot and forwarded to the Detector.

Dynamic obstacles computed by the Detector are sent to the Copilot via
SocketIO messages, then forwarded to mcu-firmware via Protobuf messages
to compute avoidance path, and to Monitor/Dashboards via SocketIO messages
for diplay.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>