tools: planner: introduce actions framework
The new actions framework allows to choose the camp and select a strategy.
A strategy is composed of a set of actions containing a set of poses.
Actions and poses have callback functions to execute commands before and
after each action and each pose.
These callbacks can, for example, send messages to the robots, or generate
new actions and new poses.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>