tools: add dashboard_beacon tool
The Beacon Dashboard is a web site displayed on the touchscreens
embedded the central beacon. It provides a menu and a dashboard to monitor all robots,
and give access to the dashboard of each robot.

The web server can also be accessed from any devices (PC, smartphones)
connected to the same network.

The web server listen on port `8080`.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>