tools: planner: adapt obstacle creation in case of PAMIs
In case of PAMI, the detected obstacle is at the front the real obstacle
instead of at its center.
Since we use a specific avoidance strategy that only needs to know the path
is intersecting the obstacle, the radius can be reduced to the minimum
to create a bounding box.

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>