replace pip by uv
uv is "an extremely fast Python package and project manager,
written in Rust" [1].

Some of its features are particularly interesting for COGIP tools:

* Very fast packages installation
Allow much faster installation on Raspberry PI.

* Python installation independent from the OS
Easier to select a Python version (3.12 currently) to have the same
versions on the development PC and on the Raspberry Pi, independently from
the OS installed on each platform.
The python version is specified in the .python-version file.

* Very good cache management
Allow to deploy COGIP tools on Raspberry Pi without internet connection
if dependencies are already in the cache.
This implies removal of the devpi server previously installed
on the beacon to cache dependencies.

* Ability to install only requirements from a pyproject file
Having separate requirements.txt files is not required anymore,
so dependencies are now included directly in pyproject.toml.

* New feature to declare indexes and pin packages to a specific index
Very useful to force pre-compiled ydlidar package (and only this one)
to be installed from our github repository.

Dockerfiles have been re-written to take advantage of uv features to
improve docker image layers caching to speed-up build phases.

[1] https://docs.astral.sh/uv/

Signed-off-by: Eric Courtois <eric.courtois@gmail.com>